# .rifle/openapi.lock.yaml
openapi: 3.1.0
info:
  title: EIFD/TIF Increment Batch API
  version: 1.0.0
  description: |
    Starts and observes the nine-step EIFD/TIF increment batch. The service uses
    preloaded shared-live assessment, tax-code, agency-reference, frozen-valuation,
    and agency-equalized-valuation data. Requests never identify local files.
    Every launch pins the business clock for the complete chain. Runs share the
    live indexed stores and are not isolated snapshots.
servers:
- url: http://localhost:8080
  description: Local development
...
security: []
paths:
...
  /api/tax-rate-input-preparation-runs:
    post:
      tags:
      - TaxRateInputPreparationRuns
      summary: Start a tax-rate input preparation run
      description: |
        Creates an asynchronous run against the server's preloaded shared data.
        The response describes the accepted run resource; poll the GET endpoint
        for status, reconciliation counts, messages, outputs, and the final return
        code. Idempotency applies to this start request, not to an isolated copy of
        mutable batch data.
      operationId: startTaxRateInputPreparationRun
      x-business-rules:
      - id: clrtm751-001
        summary: Equalized-value input is nondecreasing and division input is strictly increasing by volume and property; violations set return code 16.
        observedThrough: GET /api/tax-rate-input-preparation-runs/{id} exposes FAILED status, returnCode, and rule-linked messages.
      - id: clrtm751-002
        summary: A matching division-master row stamps its division number while preserving other equalized-value fields.
      - id: clrtm751-003
        summary: An unmatched parcel uses its property number as division number, subject to the unchecked 15-to-14 digit narrowing boundary.
      - id: clrtm751-004
        summary: Division stamping reports read, written, stamped, and unmatched counts; the legacy unmatched-division counter remains zero because it is never incremented.
        observedThrough: GET /api/tax-rate-input-preparation-runs/{id} exposes divisionStamping reconciliation counts.
      - id: clrtm752-001
        summary: Divided equalized-value input is strictly increasing by town, volume, property, and tax type, with an unchecked 15-to-14 digit property-key narrowing boundary.
        observedThrough: GET /api/tax-rate-input-preparation-runs/{id} exposes FAILED status, returnCode, and rule-linked messages.
      - id: clrtm752-002
        summary: A matched tax code carries its existing composite rate and all 40 agency slots into the parcel assessment record; no rate is calculated.
      - id: clrtm752-003
        summary: A missing tax code increments the unmatched count, emits a report message, writes no agency-assessment record for that parcel, and does not fail the run.
        observedThrough: GET /api/tax-rate-input-preparation-runs/{id} exposes assessmentRecordsUnmatched, zero-or-more agency-assessment output records, and rule-linked messages.
      - id: clrtm752-004
        summary: Attachment reconciliation reports assessment records read, written, and unmatched; written plus unmatched equals read only on normal non-error processing.
        observedThrough: GET /api/tax-rate-input-preparation-runs/{id} exposes agencyAttachment reconciliation, including retained partial output.
      - id: clrtm753-001
        summary: Prior- and current-year masters are nondecreasing by division; a descending record sets return code 16 but an active group can still produce partial output.
        observedThrough: GET /api/tax-rate-input-preparation-runs/{id} exposes FAILED status, returnCode, counts, and retained outputs.
      - id: clrtm753-002
        summary: A matched division totals each year's EAV and deduplicates nonzero agencies into separate 40-entry tables, silently dropping additional unique agencies.
      - id: clrtm753-003
        summary: Prior-only agencies become D segments, current-only agencies become A segments, and unchanged agencies produce no segment.
      - id: clrtm753-004
        summary: A division present in only one year is reported and skipped without producing a comparison record for that occurrence.
        observedThrough: GET /api/tax-rate-input-preparation-runs/{id} exposes priorOnlyDivisions, currentOnlyDivisions, and rule-linked messages.
      - id: clrtm753-005
        summary: Every matched division carries 100.0 percent because the ratio calculation is inactive; this is a known legacy defect, not an approved proportional allocation formula.
      - id: clrtm753-006
        summary: Each matched division attempts one comparison record, including zero-segment records; the 50-segment output capacity is unguarded despite an 80-segment theoretical difference.
        observedThrough: GET /api/tax-rate-input-preparation-runs/{id} exposes comparison records and segment counts, including retained partial output.
      - id: clrtm755-001
        summary: D segments load the disconnect table, A segments load the annex table, unknown types load neither, and the declared input contains at most 50 total segments.
      - id: clrtm755-002
        summary: Disconnects match prior-year parcels and annexations match current-year parcels, but only classified slots 1 through 40 are searched and duplicate parcel slots post repeatedly.
      - id: clrtm755-003
        summary: Each posting is source parcel EAV times percentage divided by 100 and rounded to an integer; the current 100.0 field posts the full parcel EAV.
      - id: clrtm755-004
        summary: An existing frozen-agency key accumulates disconnected or annexed EAV and is rewritten; later failure does not roll back earlier mutations.
        observedThrough: GET /api/tax-rate-input-preparation-runs/{id} exposes rewriteOperations, partial outputs, and final returnCode.
      - id: clrtm755-005
        summary: A missing frozen-agency key is inserted with the selected EAV and listed fields zeroed, while FA-EXPINCTX remains unspecified and prior mutations are not rolled back.
        observedThrough: GET /api/tax-rate-input-preparation-runs/{id} exposes insertOperations, partial outputs, and final returnCode.
      - id: clrtm755-006
        summary: Prior- and current-year posting inputs are nondecreasing by division; descending input sets return code 16 but active nested loops may continue.
        observedThrough: GET /api/tax-rate-input-preparation-runs/{id} exposes FAILED status, partial operation counts, and retained outputs.
      - id: clrtm755-007
        summary: The posting report lists each posted value and operation totals; rewrite and insert totals count operations rather than unique keys and remain visible after partial failure.
        observedThrough: GET /api/tax-rate-input-preparation-runs/{id} exposes frozenAgencyPosting reconciliation and report output metadata.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaxRateInputPreparationRunRequest'
            example:
              businessDate: 2025-09-15
              businessTime: 12:00:00
              idempotencyKey: clerk-agency-attachment-20250915-120000
      responses:
        '201':
          description: Run accepted and queued.
          headers:
            Location:
              description: URI of the asynchronous run resource.
              required: true
              schema:
                type: string
                example: /api/tax-rate-input-preparation-runs/900719925474099
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaxRateInputPreparationRunResponse'
              example:
                id: 900719925474099
                status: QUEUED
                businessDate: 2025-09-15
                businessTime: 12:00:00
                outputs: []
                messages: []
        '400':
          description: The start request is malformed or incomplete.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaxRateInputPreparationErrorResponse'
              example:
                error: INVALID_REQUEST
                message: businessDate is required
        '409':
          description: The idempotency key was already used with different request values.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaxRateInputPreparationErrorResponse'
              example:
                error: IDEMPOTENCY_CONFLICT
                message: The idempotency key already identifies a run with different request values
  /api/tax-rate-input-preparation-runs/{id}:
    get:
      tags:
      - TaxRateInputPreparationRuns
      summary: Get a tax-rate input preparation run
      description: |
        Returns operational status and, as processing advances, the run's counts,
        outputs, messages, and final return code. Business rules are defined by the
        POST operation; this operation only observes their outcome.
      operationId: getTaxRateInputPreparationRun
      parameters:
      - name: id
        in: path
        required: true
        description: Positive server-generated surrogate run identifier.
        schema:
          type: integer
          format: int64
          minimum: 1
      responses:
        '200':
          description: Current run state and any available result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaxRateInputPreparationRunResponse'
              examples:
                reviewedUnmatchedTaxCodes:
                  summary: Reviewed evidence; error-path reconciliation, not positive-path parity
                  value:
                    id: 900719925474099
                    status: COMPLETED
                    businessDate: 2025-09-15
                    businessTime: 12:00:00
                    returnCode: 0
                    recordsRead: 30
                    recordsWritten: 10
                    recordsUpdated: 10
                    recordsRejected: 10
                    outputs:
                    - name: agencyAssessment
                      kind: AGENCY_ASSESSMENT
                      recordCount: 0
                      available: true
                    - name: attachmentReport
                      kind: REPORT
                      recordCount: 19
                      available: true
                    messages:
                    - severity: WARNING
                      code: UNMATCHED_TAX_CODE
                      text: Ten assessment records had no matching tax-code master record; no agency-assessment records were produced.
                      ruleId: clrtm752-003
                    reconciliation:
                      divisionStamping:
                        equalizedValueRecordsRead: 10
                        divisionRecordsRead: 10
                        outputRecordsWritten: 10
                        recordsStamped: 10
                        divisionRecordsUnmatched: 0
                      agencyAttachment:
                        assessmentRecordsRead: 10
                        assessmentRecordsWritten: 0
                        assessmentRecordsUnmatched: 10
                        normalCompletionBalanced: true
        '404':
          description: No run exists for the supplied surrogate identifier.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaxRateInputPreparationErrorResponse'
              example:
                error: RUN_NOT_FOUND
                message: No run exists for the supplied identifier
...
components:
  schemas:
...
    TaxRateInputPreparationRunRequest:
      type: object
      additionalProperties: false
      required:
      - businessDate
      - businessTime
      - idempotencyKey
      properties:
        businessDate:
          type: string
          format: date
          description: Business date pinned for every step of this run.
        businessTime:
          type: string
          pattern: ^(?:[01][0-9]|2[0-3]):[0-5][0-9]:[0-5][0-9]$
          description: Business time pinned for every step of this run.
        idempotencyKey:
          type: string
          minLength: 1
          maxLength: 128
          description: Caller-selected key for safely retrying the same start request.
    TaxRateInputPreparationRunResponse:
      type: object
      additionalProperties: false
      required:
      - id
      - status
      - businessDate
      - businessTime
      - outputs
      - messages
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
          description: Positive server-generated surrogate run identifier.
        status:
          type: string
          enum:
          - QUEUED
          - RUNNING
          - COMPLETED
          - FAILED
        businessDate:
          type: string
          format: date
        businessTime:
          type: string
          pattern: ^(?:[01][0-9]|2[0-3]):[0-5][0-9]:[0-5][0-9]$
        returnCode:
          type: integer
          minimum: 0
          description: Final batch return code; omitted until one is available.
        recordsRead:
          type: integer
          minimum: 0
          description: Sum of read operations reported by completed batch steps; stage counts remain authoritative for reconciliation.
        recordsWritten:
          type: integer
          minimum: 0
          description: Sum of successful write and insert operations reported by completed batch steps, including retained partial output.
        recordsUpdated:
          type: integer
          minimum: 0
          description: Sum of successful stamp and rewrite operations reported by completed batch steps.
        recordsRejected:
          type: integer
          minimum: 0
          description: Count of reported unmatched or otherwise omitted input records; these need not fail the run.
        outputs:
          type: array
          items:
            $ref: '#/components/schemas/TaxRateInputPreparationOutput'
        messages:
          type: array
          items:
            $ref: '#/components/schemas/TaxRateInputPreparationMessage'
        reconciliation:
          $ref: '#/components/schemas/TaxRateInputPreparationReconciliation'
    TaxRateInputPreparationOutput:
      type: object
      additionalProperties: false
      required:
      - name
      - kind
      - recordCount
      - available
      properties:
        name:
          type: string
          description: Stable logical output name; never a local filesystem path.
        kind:
          type: string
          enum:
          - DIVIDED_VALUE
          - AGENCY_ASSESSMENT
          - ANNEX_DISCONNECT
          - FROZEN_AGENCY
          - REPORT
        recordCount:
          type: integer
          minimum: 0
        available:
          type: boolean
          description: Whether this server-managed output can currently be retrieved by the deployment's output facility.
    TaxRateInputPreparationMessage:
      type: object
      additionalProperties: false
      required:
      - severity
      - code
      - text
      properties:
        severity:
          type: string
          enum:
          - INFO
          - WARNING
          - ERROR
        code:
          type: string
        text:
          type: string
        ruleId:
          type: string
          description: Exact accepted rule ID; required for business-rule messages and omitted for operational worker or file I/O messages.
          enum:
          - clrtm751-001
          - clrtm751-002
          - clrtm751-003
          - clrtm751-004
          - clrtm752-001
          - clrtm752-002
          - clrtm752-003
          - clrtm752-004
          - clrtm753-001
          - clrtm753-002
          - clrtm753-003
          - clrtm753-004
          - clrtm753-005
          - clrtm753-006
          - clrtm755-001
          - clrtm755-002
          - clrtm755-003
          - clrtm755-004
          - clrtm755-005
          - clrtm755-006
          - clrtm755-007
    TaxRateInputPreparationReconciliation:
      type: object
      additionalProperties: false
      properties:
        divisionStamping:
          $ref: '#/components/schemas/TaxRateInputPreparationDivisionStampingCounts'
        agencyAttachment:
          $ref: '#/components/schemas/TaxRateInputPreparationAgencyAttachmentCounts'
        agencyComparison:
          $ref: '#/components/schemas/TaxRateInputPreparationAgencyComparisonCounts'
        frozenAgencyPosting:
          $ref: '#/components/schemas/TaxRateInputPreparationFrozenAgencyPostingCounts'
    TaxRateInputPreparationDivisionStampingCounts:
      type: object
      additionalProperties: false
      required:
      - equalizedValueRecordsRead
      - divisionRecordsRead
      - outputRecordsWritten
      - recordsStamped
      - divisionRecordsUnmatched
      properties:
        equalizedValueRecordsRead:
          type: integer
          minimum: 0
        divisionRecordsRead:
          type: integer
          minimum: 0
        outputRecordsWritten:
          type: integer
          minimum: 0
        recordsStamped:
          type: integer
          minimum: 0
        divisionRecordsUnmatched:
          type: integer
          minimum: 0
          description: Legacy displayed counter; source evidence shows it is never incremented and therefore is not proof that all division rows matched.
    TaxRateInputPreparationAgencyAttachmentCounts:
      type: object
      additionalProperties: false
      required:
      - assessmentRecordsRead
      - assessmentRecordsWritten
      - assessmentRecordsUnmatched
      - normalCompletionBalanced
      properties:
        assessmentRecordsRead:
          type: integer
          minimum: 0
        assessmentRecordsWritten:
          type: integer
          minimum: 0
          description: Agency-assessment records produced; the reviewed evidence value is zero.
        assessmentRecordsUnmatched:
          type: integer
          minimum: 0
          description: Tax-code misses omitted from agency-assessment output; the reviewed evidence value is ten.
        normalCompletionBalanced:
          type: boolean
          description: Whether written plus unmatched equals read; guaranteed only for normal non-error processing.
    TaxRateInputPreparationAgencyComparisonCounts:
      type: object
      additionalProperties: false
      required:
      - priorRecordsRead
      - currentRecordsRead
      - comparisonRecordsWritten
      - disconnectSegments
      - annexSegments
      - priorOnlyDivisions
      - currentOnlyDivisions
      properties:
        priorRecordsRead:
          type: integer
          minimum: 0
        currentRecordsRead:
          type: integer
          minimum: 0
        comparisonRecordsWritten:
          type: integer
          minimum: 0
        disconnectSegments:
          type: integer
          minimum: 0
        annexSegments:
          type: integer
          minimum: 0
        priorOnlyDivisions:
          type: integer
          minimum: 0
          description: Prior-year division occurrences reported without a current-year match.
        currentOnlyDivisions:
          type: integer
          minimum: 0
          description: Current-year division occurrences reported without a prior-year match.
    TaxRateInputPreparationFrozenAgencyPostingCounts:
      type: object
      additionalProperties: false
      required:
      - priorRecordsRead
      - currentRecordsRead
      - comparisonRecordsRead
      - rewriteOperations
      - insertOperations
      properties:
        priorRecordsRead:
          type: integer
          minimum: 0
        currentRecordsRead:
          type: integer
          minimum: 0
        comparisonRecordsRead:
          type: integer
          minimum: 0
        rewriteOperations:
          type: integer
          minimum: 0
          description: Successful rewrite operations, not unique frozen-agency keys.
        insertOperations:
          type: integer
          minimum: 0
          description: Successful insert operations, not unique frozen-agency keys.
    TaxRateInputPreparationErrorResponse:
      type: object
      additionalProperties: false
      required:
      - error
      - message
      properties:
        error:
          type: string
        message:
          type: string
        ruleId:
          type: string
          description: Exact accepted business-rule identifier associated with the error context.
...